pyproject.toml
[tool.poetry]
name = "test-lingvanex-1"
version = "0.1.0"
description = ""
authors = ["polenom <vitalimitsevich@yandex.by>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
psycopg2 = "^2.9.5"
aiohttp = "^3.8.3"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
